(kernel_stat): Expand time_t to 'long int' not, 'unsigned int'.  Use 'long int' for __unused? members.

+++ sysdeps/unix/sysv/linux/mips/kernel_stat.h
@@ -12,12 +12,12 @@ struct kernel_stat
     long int __pad2[2];
     long int st_size;
     long int __pad3;
-    unsigned int st_atime;
-    unsigned int __unused1;
-    unsigned int st_mtime;
-    unsigned int __unused2;
-    unsigned int st_ctime;
-    unsigned int __unused3;
+    long int st_atime;
+    long int __unused1;
+    long int st_mtime;
+    long int __unused2;
+    long int st_ctime;
+    long int __unused3;
     long int st_blksize;
     long int st_blocks;
     char st_fstype[16];			/* Filesystem type name, unsupported */
